+++ weblate/permissions/templatetags/permissions.py
@@ -103,11 +103,6 @@ def can_reset_translation(user, project):
 
 
 @register.simple_tag
-def can_lock_translation(user, project):
-    return weblate.permissions.helpers.can_lock_translation(user, project)
-
-
-@register.simple_tag
 def can_lock_subproject(user, project):
     return weblate.permissions.helpers.can_lock_subproject(user, project)