+++ weblate/trans/views/lock.py
@@ -29,61 +29,7 @@ from weblate.trans.util import redirect_param
 from weblate.trans.views.helper import (
     get_project, get_subproject, get_translation
 )
-from weblate.permissions.helpers import (
-    can_lock_subproject, can_lock_translation
-)
-
-
-@require_POST
-@login_required
-def update_lock(request, project, subproject, lang):
-    obj = get_translation(request, project, subproject, lang)
-
-    if obj.update_lock(request.user):
-        return JsonResponse(
-            data={'status': True}
-        )
-
-    response = {
-        'status': False,
-        'message': _('Failed to update lock, probably session has expired.'),
-    }
-
-    return JsonResponse(data=response)
-
-
-@require_POST
-@login_required
-def lock_translation(request, project, subproject, lang):
-    obj = get_translation(request, project, subproject, lang)
-
-    if not can_lock_translation(request.user, obj.subproject.project):
-        raise PermissionDenied()
-
-    if not obj.is_user_locked(request.user):
-        obj.create_lock(request.user, True)
-        messages.success(request, _('Translation is now locked for you.'))
-
-    return redirect_param(obj, '#locking')
-
-
-@require_POST
-@login_required
-def unlock_translation(request, project, subproject, lang):
-    obj = get_translation(request, project, subproject, lang)
-
-    if not can_lock_translation(request.user, obj.subproject.project):
-        raise PermissionDenied()
-
-    if not obj.is_user_locked(request.user):
-        obj.create_lock(None)
-        messages.success(
-            request,
-            _('Translation is now open for translation updates.')
-        )
-
-    return redirect_param(obj, '#locking')
-
+from weblate.permissions.helpers import can_lock_subproject
 
 @require_POST
 @login_required